move utility scripts to template files
